dia: Atlantic salmon Dam Impact Analysis (DIA) for R

The dia package is a collection of functions and datasets used to run the Dam Impact Analysis (DIA) for Atlantic salmon in the Penobscot River, Maine, USA.

Details

Additional examples and descriptions of the primary user-facing functions (run_dia() and run_dia_shiny()) can be found on the GitHub repository for the `dia` package: https://github.com/danStich/dia.

Functions called directly

The following functions can be called directly from dia.

run_dia The primary user-facing function used to run the multi-generation DIA model for Atlantic salmon run_dia_shiny A `shiny` implementation of `run_dia()` with graphical user interface run_one_gen Run the DIA model for a single generation get_dam_passage Get flow-correlated downstream dam passage values and probability of using Stillwater Branch for smolts each year in `run_one_gen` make_egg_to_smolt_survival Simulate survival from egg to smolt stage for naturally produced smolts in `run_one_gen` make_eggs_per_female Simulate number of eggs per female each year in `run_one_gen` make_inriver_s Simulate natural, in-river survival per km for downstream migrating smolts in `run_one_gen` make_marine_s Simulate marine survival each year in `run_one_gen` run_downstream_passage Run the downstream smolt passage module in `run_one_gen` run_inefficiency Re-distribute adult upstream migrants following failed dam passage in `run_one_gen` run_straying Assign destination production units for adult returns in `run_one_gen` run_upstream_passage Function used to run the upstream passage module for adults in `run_one_gen` smolt_stocking Function used to distribute stocked smolts among production units in `run_one_gen`

Data

The following built-in datasets are included:

downstream_brownsmills Flow-correlated estimates of smolt survival through Browns Mills Dam downstream_frankfort Flow-correlated estimates of smolt survival through Frankfort Dam downstream_greatworks Flow-correlated estimates of smolt survival through Great Works Dam downstream_howland Flow-correlated estimates of smolt survival through Howland Dam downstream_lowell Flow-correlated estimates of smolt survival through Lowell Dam downstream_mattaceunk Flow-correlated estimates of smolt survival through Mattaceunk Dam downstream_medway Flow-correlated estimates of smolt survival through Medway Dam downstream_milford Updated (2020) flow-correlated estimates of smolt survival through Milford Dam downstream_milford_old Original (2013) flow-correlated estimates of smolt survival through Milford Dam downstream_milo Flow-correlated estimates of smolt survival through Milo Dam downstream_moosehead Flow-correlated estimates of smolt survival through Moosehead Dam downstream_orono Updated (2020) flow-correlated estimates of smolt survival through Orono Dam downstream_orono_old Original (2013) flow-correlated estimates of smolt survival through Orono Dam downstream_sebec Flow-correlated estimates of smolt survival through Sebec Dam downstream_stillwater Updated (2020) flow-correlated estimates of smolt survival through Stillwater Dam downstream_stillwater_old Original (2013) flow-correlated estimates of smolt survival through Stillwater Dam downstream_veazie Flow-correlated estimates of smolt survival through Veazie Dam downstream_westenfield Flow-correlated estimates of smolt survival through West Enfield Dam in_river_m Cumulative frequency distribution of in-river mortality per km for smolts life_stage_survival Distributional parameters for year-specific number of eggs produced per female, egg-to-smolt survival rate, in-river mortality rate, and marine survival rate n_dams Number of dams passed by smolts in each production unit and migration route production_units Division of Penobscot River watershed into sections, or Production Units (PUs) stillwater_use Updated (2020) probability of smolts using Stillwater Branch for outmigration given discharge stillwater_use_old Original (2013) probability of smolts using Stillwater Branch for outmigration given discharge stocking_default Default stocking proportions of smolts by PU straying_locations Straying locations of adult Atlantic salmon based on behavioral patterns straying_matrix Numeric columns only from `straying_locations` upstream_inefficiency Destination PUs and probabilities for adults that fail to pass dams inefficiency_matrix Columns 5 through 20 from `upstream_inefficiency`