Get dam-specific survival rates

Get correlated dam-specific downstream survival rates for smolts at each dam from built-in `downstream_` datasets. Values based on Amaral et al. (2012) and sampled following Nieland et al. (2013, 2015) and Nieland and Sheehan (2020).

Usage

get_dam_passage(stoch_val = 0.1695, new_or_old = "new")

Arguments

stoch_val

Value to be used for a stochastic random draw from `runif(-stoch_val, stoch_val)` to enforce variability on quantiles used for flow-correlated survival values in the underlying data sets following Nieland and Sheehan (2020).

new_or_old

A character string indicating whether to use `"new"` (Nieland and Sheehan 2020) or `"old"` (Amaral et al. 2012, Nieland et al. 2013) flow-correlated probabilities of p_stillwater as well as flow-correlated survival at `milford`, `orono`, and `stillwater` dams.

Value

A list two elements. The first `dam_survival` returns downstream dam passage survival for smolts at each of 15 dams in the watershed. The second, `p_stillwater`, contains simulated value for probability of fish using the Stillwater Branch during downstream migration.

References

Amaral S, Fay C, Hecker G, Perkins N. 2012. Atlantic salmon survival estimates at mainstem hydroelectric projects on the Penobscot River. Phase 3 Final Report. Alden Research Laboratory, Inc., Holden, MA.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams on diadromous fish: a case study for Atlantic salmon in the Penobscot River, Maine. ICES Journal of Marine Science 72:2423–2437.

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US Department of Commerce, Northeast Fisheries Science Center Reference Document 13-09, Woods Hole, MA.

Examples

# 1. ----
# Simulate dam survival at each of 15 dams, along with p_stillwater
# using estimates from Nieland et al. (2015, 2020)
get_dam_passage(stoch_val = 0.1695, new_or_old = "new")
#> $dam_survival
#>  [1] 0.9001 0.8228 0.9241 0.8881 0.8692 0.8849 0.8870 0.9158 0.8723 0.9076
#> [11] 0.9182 0.8611 0.8904 0.9000 0.9081
#> 
#> $p_stillwater
#> [1] 0.2229
#> 

# 2. ----
# Simulate dam survival at each of 15 dams, along with p_stillwater
# using estimates from Nieland et al. (2013)
get_dam_passage(stoch_val = 0.1695, new_or_old = "old")
#> $dam_survival
#>  [1] 0.8838 0.7719 0.9249 0.8472 0.7467 0.8561 0.8522 0.9091 0.8472 0.9162
#> [11] 0.9131 0.8086 0.8831 0.8809 0.9438
#> 
#> $p_stillwater
#> [1] 0.1381
#>