Create natural in-river smolt survival by reach

Create natural in-river smolt mortality per kilometer from built-in datasets.

Usage

make_inriver_s(m = NULL, initial = FALSE, hatchery_wild = "wild", alt = 0)

Arguments

m

Natural in-river mortality rate per km. The default (`NULL`) uses a truncated normal distribution with mean = 0.00329 and standard deviation of 0.00459 defined on the interval (0, 0.29) used by Nieland and Sheehan (2020) with empirical estimates from Holbrook et al. (2011) and Stich et al. (2015a)

initial

Logical indicating Whether in-river mortality rate is for initial PU or a PU downstream from where smolts started. If `TRUE` then default hatchery segment length is used to calculate whole-reach survival for hatchery smolts as (1 - m) ^ dia::production_units$Hatchery_segment_length) and 1/2 of the longest straight-line reach length (dia::production_units$Longest_segment_length_km) is used to calculate whole-reach survival of wild smolts. If `FALSE`, then both hatchery and wild smolt survival is calculated from the longets straight-line reach length.

hatchery_wild

A character string indicating whether reach mortality is to be calculated for `"hatchery"` or `"wild"` smolts.

alt

A numeric indicating whether to use `"Longest_segment_length"` (`alt = 0`), `"Partial_segment_length_1"` (`alt = 2`), or `"Partial_segment_length_2"` (`alt = 3`) from the `production_units` table to calculate whole-reach survival rates based on per-km mortality.

Value

A numeric vector with 15 elements corresponding to PUs in the Penobscot River watershed.

References

Holbrook CM, Kinnison MT, Zydlewski J. 2011. Survival of migrating Atlantic salmon smolts through the Penobscot River, Maine: a prerestoration assessment. Transactions of the American Fisheries Society 140:1255–1268.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods Hole, MA.

Stich DS, Bailey MM, Holbrook CM, Kinnison MT, Zydlewski JD. 2015a. Catchment-wide survival of wild- and hatchery-reared Atlantic salmon smolts in a changing system. Canadian Journal of Fisheries and Aquatic Sciences 72:1352–1365.

Examples

# 1. Simulate in-river survival value for wild fish in each production unit ----
make_inriver_s(hatchery_wild = "wild")
#>  [1] 0.3951882 0.6590039 0.5352947 0.7910835 0.9718607 0.8375569 0.9634874
#>  [8] 0.6869020 0.8631033 0.8337657 0.9920697 0.9883211 0.8631033 0.8502342
#> [15] 0.6952070

# 2. Simulate in-river survival value for hatchery fish in each production unit ----
make_inriver_s(hatchery_wild = "hatchery")
#>  [1] 0.08358083 0.32795941 0.18811690 0.53445412 0.92653299 0.62257067
#>  [7] 0.90534582 0.36640300 0.67463976 0.61506538 0.97893968 0.96908225
#> [13] 0.67463976 0.64808250 0.37836620

# 3. Simulate post-stocking survival value for hatchery fish in each production unit ----
make_inriver_s(initial = TRUE, hatchery_wild = "hatchery")
#>  [1] 0.9445699 0.9647330 0.8838877 0.9775713 0.9970000 0.9377833 0.9634874
#>  [8] 0.8733549 0.9806896 0.9010483 0.9950548 0.9885884 0.9690615 0.9873713
#> [15] 0.8318141