Run the Dam Impact Analysis

Use functions from dia to simulate multiple generations of 2 sea-winter female Atlantic salmon through time using data and inputs from Nieland et al. (2013, 2015) and Nieland and Sheehan (2020) as implemented in `@Risk` add-in for `Excel`. This function is a wrapper for run_one_gen that uses output from one generation as the seed for user-specified number of subsequent generations.

Usage

run_dia(
  n_generations = 15,
  n_wild = 31,
  n_hatchery = 306,
  stocking = 1,
  n_stocked = rep(545000, 15),
  upstream = list(medway = 0, mattaceunk = 0.9, west_enfield = 0.95, upper_dover = 0.92,
    browns_mills = 0.92, sebec = 0, milo = 0, howland = 0.95, lowel = 0.92, stillwater =
    0, milford = 0.95, great_works = 1, orono = 0, veazie = 1, frankfort = 1),
  downstream = list(medway = 0, mattaceunk = 1, west_enfield = 0.96, upper_dover =
    0.9215, browns_mills = NA, sebec = NA, milo = NA, howland = 1, lowell = NA,
    stillwater = 0.96, milford = 0.96, great_works = 1, orono = 0.96, veazie = 1,
    frankfort = NA),
  in_river_s = NA,
  mattaceunk_impoundment_mortality = 0.072,
  p_stillwater = NA,
  indirect_latent_mortality = 0.06,
  p_female = 0.6,
  new_or_old = "new",
  marine_s_hatchery = NA,
  marine_s_wild = NA,
  straying_matrix = NULL,
  p_mainstem_up = 1,
  n_broodstock = 150
)

Arguments

n_generations

Number of generations to simulate. One complete generation from egg to returning adult is six years.

n_wild

Number of starting wild, adult Atlantic salmon.

n_hatchery

Number of starting hatchery adult Atlantic salmon.

stocking

Numeric indicating whether hatchery stocking is on (`1`) or off (`0`). If hatchery stocking is on, then hatchery smolts are stocked into production units based on the proportion_stocked column in the built-in stocking_default dataset.

n_stocked

A numeric vector corresponding to `n_generations` in length with the number of stocked hatchery smolts per generation.

upstream

Upstream passage efficiency of adult Atlantic salmon through dams. Default values are based on NMFS (2013) prescriptions as implemented in Nieland and Sheehan (2020). Any numeric value between zero and 1, inclusive, can be specified for any dam by the user.

downstream

Downstream survival of Atlantic salmon smolts through dams. The default values are based on the most recent values used by Nieland et al. (2020) based on standards established in the the NOAA Species Protection Plan following implementation of the Penobscot River Restoration Project. The default `NA` randomly samples correlated survival rates of smolts for each dam based on cumulative flow probabilities and associated empirical survival rates (Nieland et al. 2013, Nieland and Sheehan 2020), and can also be used for dams that have a default value other than NA.

in_river_s

In-river survival per kilometer for downstream migrating smolts. The default `NA` value simulates from a cumulative distribution function using values in in_river_m. This value is the same for hatchery stocked and wild smolts.

mattaceunk_impoundment_mortality

Mortality incurred by Atlantic salmon smolts during migration through the Mattaceunk (Weldon) Dam impoundment. The default value is based on Nieland and Sheehan (2020), who used results of Holbrook et al. (2011) and Stich et al. (2015a).

p_stillwater

Probability that fish use the Stillwater Branch for downstream migration. The default (NULL) draws flow-correlated probability of using the Stillwater Branch from a cumulative distribution of flows with paired estimates of p_stillwater used by Nieland and Sheehan (2020), based on empirical results in Holbrook et al. (2006), Stich et al. (2014), and Stich et al. (2015a).

indirect_latent_mortality

Indirect, latent mortality incurred by Atlantic salmon smolts at each dam passed. The default value of 0.06 is what was used in Nieland and Sheehan (2020), derived from estimates in Stich et al. (2015b).

p_female

Proportion of females in population. Any value between zero and one, inclusive, is allowable.

new_or_old

A character string indicating whether to use "new" (Nieland and Sheehan 2020) or "old" (Nieland et al. 2013, 2015) flow-correlated probabilities of p_stillwater as well as flow-correlated survival at milford, orono, and stillwater dams.

marine_s_hatchery

Numeric indicating marine survival rate for post-smolt to adult survival of hatchery outmigrants. The default (`NA`) simulates values from a truncated normal distribution using hatchery smolt survival estimates from the Penobscot River, ME, USA.

marine_s_wild

Numeric indicating marine survival rate for post-smolt to adult survival of wild outmigrants during two winters at sea. The default (`NA`) simulates values from a truncated normal distribution using wild smolt survival estimates from the Narraguagus River, ME, USA.

straying_matrix

A dataframe identical in structure to the built-in straying_locations dataset. By default, uses the built-in data set to parameterize straying probabilities to destination PUs from starting PUs.

p_mainstem_up

Probability that fish use the mainstem Penobscot River (and not Stillwater Branch) for upstream migration around Marsh Island. The default value `1` was used by Nieland and Sheehan (2020).

n_broodstock

Target number of adult returns collected at Milford Dam for spawning at US Fish and Wildlife Service Craig Brook National Fish Hatchery each year. Broodstock are collected upstream of Milford Dam in run_upstream_passage and removed from the population for that generation, as is the case in reality.

Value

A dataframe with four variables corresponding to each generation of the simulation. Variables correspond to output from run_one_gen.

production_unit Production unit generation Generation number from simulation origin Fish origin (hatchery or wild) abundance Fish abundance within production units, generations, and origin

References

Holbrook CM, Kinnison MT, Zydlewski J. 2011. Survival of migrating Atlantic salmon smolts through the Penobscot River, Maine: a prerestoration assessment. Transactions of the American Fisheries Society 140:1255-1268.

Holbrook CM, Zydlewski J, Gorsky D, Shepard SL, Kinnison MT. 2009. Movements of prespawn adult Atlantic salmon near hydroelectric dams in the lower Penobscot River, Maine. North American Journal of Fisheries Management 29:495-505.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams on diadromous fish: a case study for Atlantic salmon in the Penobscot River, Maine. ICES Journal of Marine Science 72:2423-2437.

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US Department of Commerce, Northeast Fisheries Science Center Reference Document 13-09, Woods Hole, MA.

NMFS (National Marine Fisheries Service). 2013. Endangered Species Act Biological Opinion: FERC amendment of license for the Mattaceunk Project (Interim Species Protection Plan). NMFS Greater Atlantic Regional Office, NER-2013-9640, Gloucester, MA.

Stich DS, Bailey MM, Holbrook CM, Kinnison MT, Zydlewski JD. 2015a. Catchment-wide survival of wild- and hatchery-reared Atlantic salmon smolts in a changing system. Canadian Journal of Fisheries and Aquatic Sciences 72:1352-1365.

Stich DS, Bailey MM, Zydlewski JD. 2014. Survival of Atlantic salmon Salmo salar smolts through a hydropower complex. Journal of Fish Biology 85:1074-1096.

Stich DS, Kinnison MT, Kocik JF, Zydlewski JD. 2015b. Initiation of migration and movement rates of Atlantic salmon smolts in fresh water. Canadian Journal of Fisheries and Aquatic Sciences 72:1339-1351.

Examples

# Additional examples, including parallel implementation can be found in the
# readme file for the GitHub repository for this package: 
# https://github.com/danStich/dia.

# 1. Run dia model with defaults for 15 generations ----

result <- run_dia(n_generations = 15)

# 2. Run dia model with user-specified dam passage ----

run_dia(
  n_generations = 15,
  upstream = list(
    medway = 0, 
    mattaceunk = 0.9, 
    west_enfield = 0.95, 
    upper_dover = 0.92, 
    browns_mills = 0.92, 
    sebec = 0, 
    milo = 0, 
    howland = 0.95, 
    lowel = 0.92,
    stillwater = 0, 
    milford = 0.95, 
    great_works = 1, 
    orono = 0, 
    veazie = 1, 
    frankfort = 1),
  downstream = list(
  medway = 0, 
  mattaceunk = 1, 
  west_enfield = 0.96, 
  upper_dover = 0.9215, 
  browns_mills = 1, 
  sebec = 1, 
  milo = 1, 
  howland = 1, 
  lowell = 1,
  stillwater = 0.96, 
  milford = 0.96, 
  great_works = 1, 
  orono = 0.96, 
  veazie = 1,
  frankfort = 1))
#>     production_unit generation   origin abundance
#> 1             PU_01          1 hatchery         0
#> 2             PU_01          1     wild         0
#> 3             PU_02          1 hatchery         0
#> 4             PU_02          1     wild         0
#> 5             PU_03          1 hatchery         0
#> 6             PU_03          1     wild         0
#> 7             PU_06          1 hatchery         0
#> 8             PU_06          1     wild         0
#> 9             PU_05          1 hatchery         0
#> 10            PU_05          1     wild         0
#> 11            PU_08          1 hatchery         0
#> 12            PU_08          1     wild         0
#> 13            PU_07          1 hatchery         0
#> 14            PU_07          1     wild         0
#> 15            PU_04          1 hatchery         0
#> 16            PU_04          1     wild         0
#> 17            PU_15          1 hatchery         0
#> 18            PU_15          1     wild         0
#> 19            PU_09          1 hatchery        42
#> 20            PU_09          1     wild        14
#> 21            PU_10          1 hatchery         0
#> 22            PU_10          1     wild         0
#> 23            PU_11          1 hatchery         4
#> 24            PU_11          1     wild         1
#> 25            PU_12          1 hatchery         0
#> 26            PU_12          1     wild         0
#> 27            PU_13          1 hatchery         1
#> 28            PU_13          1     wild         7
#> 29            PU_14          1 hatchery       211
#> 30            PU_14          1     wild        77
#> 31            PU_01          2 hatchery         0
#> 32            PU_01          2     wild         0
#> 33            PU_02          2 hatchery         0
#> 34            PU_02          2     wild         0
#> 35            PU_03          2 hatchery         4
#> 36            PU_03          2     wild         1
#> 37            PU_06          2 hatchery         2
#> 38            PU_06          2     wild         0
#> 39            PU_05          2 hatchery         0
#> 40            PU_05          2     wild         0
#> 41            PU_08          2 hatchery         0
#> 42            PU_08          2     wild         0
#> 43            PU_07          2 hatchery         0
#> 44            PU_07          2     wild         0
#> 45            PU_04          2 hatchery        22
#> 46            PU_04          2     wild         2
#> 47            PU_15          2 hatchery         3
#> 48            PU_15          2     wild         0
#> 49            PU_09          2 hatchery       168
#> 50            PU_09          2     wild        35
#> 51            PU_10          2 hatchery         0
#> 52            PU_10          2     wild         0
#> 53            PU_11          2 hatchery         9
#> 54            PU_11          2     wild         3
#> 55            PU_12          2 hatchery         0
#> 56            PU_12          2     wild         0
#> 57            PU_13          2 hatchery         3
#> 58            PU_13          2     wild         3
#> 59            PU_14          2 hatchery       639
#> 60            PU_14          2     wild       158
#> 61            PU_01          3 hatchery         0
#> 62            PU_01          3     wild         0
#> 63            PU_02          3 hatchery         1
#> 64            PU_02          3     wild         1
#> 65            PU_03          3 hatchery        18
#> 66            PU_03          3     wild        24
#> 67            PU_06          3 hatchery         8
#> 68            PU_06          3     wild         1
#> 69            PU_05          3 hatchery         0
#> 70            PU_05          3     wild         0
#> 71            PU_08          3 hatchery         0
#> 72            PU_08          3     wild         0
#> 73            PU_07          3 hatchery         0
#> 74            PU_07          3     wild         0
#> 75            PU_04          3 hatchery       109
#> 76            PU_04          3     wild        51
#> 77            PU_15          3 hatchery        13
#> 78            PU_15          3     wild        11
#> 79            PU_09          3 hatchery       226
#> 80            PU_09          3     wild       247
#> 81            PU_10          3 hatchery         0
#> 82            PU_10          3     wild         0
#> 83            PU_11          3 hatchery        10
#> 84            PU_11          3     wild        15
#> 85            PU_12          3 hatchery         0
#> 86            PU_12          3     wild         0
#> 87            PU_13          3 hatchery         4
#> 88            PU_13          3     wild        11
#> 89            PU_14          3 hatchery       830
#> 90            PU_14          3     wild       973
#> 91            PU_01          4 hatchery         0
#> 92            PU_01          4     wild         0
#> 93            PU_02          4 hatchery         0
#> 94            PU_02          4     wild         0
#> 95            PU_03          4 hatchery         2
#> 96            PU_03          4     wild         0
#> 97            PU_06          4 hatchery         1
#> 98            PU_06          4     wild         0
#> 99            PU_05          4 hatchery         0
#> 100           PU_05          4     wild         0
#> 101           PU_08          4 hatchery         0
#> 102           PU_08          4     wild         0
#> 103           PU_07          4 hatchery         0
#> 104           PU_07          4     wild         0
#> 105           PU_04          4 hatchery         9
#> 106           PU_04          4     wild         1
#> 107           PU_15          4 hatchery         1
#> 108           PU_15          4     wild         0
#> 109           PU_09          4 hatchery       141
#> 110           PU_09          4     wild        17
#> 111           PU_10          4 hatchery         0
#> 112           PU_10          4     wild         0
#> 113           PU_11          4 hatchery         7
#> 114           PU_11          4     wild         1
#> 115           PU_12          4 hatchery         0
#> 116           PU_12          4     wild         0
#> 117           PU_13          4 hatchery         2
#> 118           PU_13          4     wild         1
#> 119           PU_14          4 hatchery       533
#> 120           PU_14          4     wild        59
#> 121           PU_01          5 hatchery         0
#> 122           PU_01          5     wild         0
#> 123           PU_02          5 hatchery         2
#> 124           PU_02          5     wild         0
#> 125           PU_03          5 hatchery        28
#> 126           PU_03          5     wild         6
#> 127           PU_06          5 hatchery        10
#> 128           PU_06          5     wild         0
#> 129           PU_05          5 hatchery         0
#> 130           PU_05          5     wild         0
#> 131           PU_08          5 hatchery         0
#> 132           PU_08          5     wild         0
#> 133           PU_07          5 hatchery         0
#> 134           PU_07          5     wild         0
#> 135           PU_04          5 hatchery       147
#> 136           PU_04          5     wild        12
#> 137           PU_15          5 hatchery        18
#> 138           PU_15          5     wild         3
#> 139           PU_09          5 hatchery       346
#> 140           PU_09          5     wild        67
#> 141           PU_10          5 hatchery         0
#> 142           PU_10          5     wild         0
#> 143           PU_11          5 hatchery        15
#> 144           PU_11          5     wild         4
#> 145           PU_12          5 hatchery         0
#> 146           PU_12          5     wild         0
#> 147           PU_13          5 hatchery         6
#> 148           PU_13          5     wild         2
#> 149           PU_14          5 hatchery      1263
#> 150           PU_14          5     wild       259
#> 151           PU_01          6 hatchery         0
#> 152           PU_01          6     wild         0
#> 153           PU_02          6 hatchery         1
#> 154           PU_02          6     wild         0
#> 155           PU_03          6 hatchery         9
#> 156           PU_03          6     wild         1
#> 157           PU_06          6 hatchery         3
#> 158           PU_06          6     wild         0
#> 159           PU_05          6 hatchery         0
#> 160           PU_05          6     wild         0
#> 161           PU_08          6 hatchery         0
#> 162           PU_08          6     wild         0
#> 163           PU_07          6 hatchery         0
#> 164           PU_07          6     wild         0
#> 165           PU_04          6 hatchery        40
#> 166           PU_04          6     wild         3
#> 167           PU_15          6 hatchery         5
#> 168           PU_15          6     wild         1
#> 169           PU_09          6 hatchery       197
#> 170           PU_09          6     wild        20
#> 171           PU_10          6 hatchery         0
#> 172           PU_10          6     wild         0
#> 173           PU_11          6 hatchery        10
#> 174           PU_11          6     wild         1
#> 175           PU_12          6 hatchery         0
#> 176           PU_12          6     wild         0
#> 177           PU_13          6 hatchery         4
#> 178           PU_13          6     wild         0
#> 179           PU_14          6 hatchery       759
#> 180           PU_14          6     wild        85
#> 181           PU_01          7 hatchery         0
#> 182           PU_01          7     wild         0
#> 183           PU_02          7 hatchery         0
#> 184           PU_02          7     wild         0
#> 185           PU_03          7 hatchery         0
#> 186           PU_03          7     wild         0
#> 187           PU_06          7 hatchery         0
#> 188           PU_06          7     wild         0
#> 189           PU_05          7 hatchery         0
#> 190           PU_05          7     wild         0
#> 191           PU_08          7 hatchery         0
#> 192           PU_08          7     wild         0
#> 193           PU_07          7 hatchery         0
#> 194           PU_07          7     wild         0
#> 195           PU_04          7 hatchery         0
#> 196           PU_04          7     wild         0
#> 197           PU_15          7 hatchery         0
#> 198           PU_15          7     wild         0
#> 199           PU_09          7 hatchery        93
#> 200           PU_09          7     wild        24
#> 201           PU_10          7 hatchery         0
#> 202           PU_10          7     wild         0
#> 203           PU_11          7 hatchery         5
#> 204           PU_11          7     wild         2
#> 205           PU_12          7 hatchery         0
#> 206           PU_12          7     wild         0
#> 207           PU_13          7 hatchery         2
#> 208           PU_13          7     wild         0
#> 209           PU_14          7 hatchery       366
#> 210           PU_14          7     wild       103
#> 211           PU_01          8 hatchery         0
#> 212           PU_01          8     wild         0
#> 213           PU_02          8 hatchery         2
#> 214           PU_02          8     wild         0
#> 215           PU_03          8 hatchery        24
#> 216           PU_03          8     wild         1
#> 217           PU_06          8 hatchery        10
#> 218           PU_06          8     wild         0
#> 219           PU_05          8 hatchery         0
#> 220           PU_05          8     wild         0
#> 221           PU_08          8 hatchery         0
#> 222           PU_08          8     wild         0
#> 223           PU_07          8 hatchery         0
#> 224           PU_07          8     wild         0
#> 225           PU_04          8 hatchery       138
#> 226           PU_04          8     wild         2
#> 227           PU_15          8 hatchery        15
#> 228           PU_15          8     wild         1
#> 229           PU_09          8 hatchery       308
#> 230           PU_09          8     wild        10
#> 231           PU_10          8 hatchery         0
#> 232           PU_10          8     wild         0
#> 233           PU_11          8 hatchery        14
#> 234           PU_11          8     wild         1
#> 235           PU_12          8 hatchery         0
#> 236           PU_12          8     wild         0
#> 237           PU_13          8 hatchery         6
#> 238           PU_13          8     wild         0
#> 239           PU_14          8 hatchery      1164
#> 240           PU_14          8     wild        37
#> 241           PU_01          9 hatchery         0
#> 242           PU_01          9     wild         0
#> 243           PU_02          9 hatchery         0
#> 244           PU_02          9     wild         1
#> 245           PU_03          9 hatchery         6
#> 246           PU_03          9     wild        15
#> 247           PU_06          9 hatchery         3
#> 248           PU_06          9     wild         3
#> 249           PU_05          9 hatchery         0
#> 250           PU_05          9     wild         0
#> 251           PU_08          9 hatchery         0
#> 252           PU_08          9     wild         0
#> 253           PU_07          9 hatchery         0
#> 254           PU_07          9     wild         0
#> 255           PU_04          9 hatchery        30
#> 256           PU_04          9     wild        54
#> 257           PU_15          9 hatchery         4
#> 258           PU_15          9     wild        14
#> 259           PU_09          9 hatchery        93
#> 260           PU_09          9     wild       247
#> 261           PU_10          9 hatchery         0
#> 262           PU_10          9     wild         0
#> 263           PU_11          9 hatchery         4
#> 264           PU_11          9     wild        12
#> 265           PU_12          9 hatchery         0
#> 266           PU_12          9     wild         0
#> 267           PU_13          9 hatchery         2
#> 268           PU_13          9     wild         8
#> 269           PU_14          9 hatchery       341
#> 270           PU_14          9     wild       896
#> 271           PU_01         10 hatchery         0
#> 272           PU_01         10     wild         0
#> 273           PU_02         10 hatchery         2
#> 274           PU_02         10     wild         0
#> 275           PU_03         10 hatchery        30
#> 276           PU_03         10     wild         1
#> 277           PU_06         10 hatchery        12
#> 278           PU_06         10     wild         0
#> 279           PU_05         10 hatchery         0
#> 280           PU_05         10     wild         0
#> 281           PU_08         10 hatchery         0
#> 282           PU_08         10     wild         0
#> 283           PU_07         10 hatchery         0
#> 284           PU_07         10     wild         0
#> 285           PU_04         10 hatchery       156
#> 286           PU_04         10     wild         2
#> 287           PU_15         10 hatchery        17
#> 288           PU_15         10     wild         1
#> 289           PU_09         10 hatchery       348
#> 290           PU_09         10     wild         8
#> 291           PU_10         10 hatchery         0
#> 292           PU_10         10     wild         0
#> 293           PU_11         10 hatchery        15
#> 294           PU_11         10     wild         1
#> 295           PU_12         10 hatchery         0
#> 296           PU_12         10     wild         0
#> 297           PU_13         10 hatchery         6
#> 298           PU_13         10     wild         0
#> 299           PU_14         10 hatchery      1296
#> 300           PU_14         10     wild        31
#> 301           PU_01         11 hatchery         0
#> 302           PU_01         11     wild         0
#> 303           PU_02         11 hatchery         0
#> 304           PU_02         11     wild         0
#> 305           PU_03         11 hatchery         0
#> 306           PU_03         11     wild         0
#> 307           PU_06         11 hatchery         0
#> 308           PU_06         11     wild         0
#> 309           PU_05         11 hatchery         0
#> 310           PU_05         11     wild         0
#> 311           PU_08         11 hatchery         0
#> 312           PU_08         11     wild         0
#> 313           PU_07         11 hatchery         0
#> 314           PU_07         11     wild         0
#> 315           PU_04         11 hatchery         0
#> 316           PU_04         11     wild         0
#> 317           PU_15         11 hatchery         0
#> 318           PU_15         11     wild         0
#> 319           PU_09         11 hatchery        63
#> 320           PU_09         11     wild        75
#> 321           PU_10         11 hatchery         0
#> 322           PU_10         11     wild         0
#> 323           PU_11         11 hatchery         4
#> 324           PU_11         11     wild         6
#> 325           PU_12         11 hatchery         0
#> 326           PU_12         11     wild         0
#> 327           PU_13         11 hatchery         1
#> 328           PU_13         11     wild         3
#> 329           PU_14         11 hatchery       255
#> 330           PU_14         11     wild       357
#> 331           PU_01         12 hatchery         0
#> 332           PU_01         12     wild         0
#> 333           PU_02         12 hatchery         0
#> 334           PU_02         12     wild         0
#> 335           PU_03         12 hatchery         4
#> 336           PU_03         12     wild         4
#> 337           PU_06         12 hatchery         1
#> 338           PU_06         12     wild         0
#> 339           PU_05         12 hatchery         0
#> 340           PU_05         12     wild         0
#> 341           PU_08         12 hatchery         0
#> 342           PU_08         12     wild         0
#> 343           PU_07         12 hatchery         0
#> 344           PU_07         12     wild         0
#> 345           PU_04         12 hatchery        16
#> 346           PU_04         12     wild         8
#> 347           PU_15         12 hatchery         2
#> 348           PU_15         12     wild         2
#> 349           PU_09         12 hatchery       139
#> 350           PU_09         12     wild       127
#> 351           PU_10         12 hatchery         0
#> 352           PU_10         12     wild         0
#> 353           PU_11         12 hatchery         7
#> 354           PU_11         12     wild         9
#> 355           PU_12         12 hatchery         0
#> 356           PU_12         12     wild         0
#> 357           PU_13         12 hatchery         2
#> 358           PU_13         12     wild         5
#> 359           PU_14         12 hatchery       515
#> 360           PU_14         12     wild       521
#> 361           PU_01         13 hatchery         0
#> 362           PU_01         13     wild         0
#> 363           PU_02         13 hatchery         4
#> 364           PU_02         13     wild         1
#> 365           PU_03         13 hatchery        48
#> 366           PU_03         13     wild        12
#> 367           PU_06         13 hatchery        22
#> 368           PU_06         13     wild         0
#> 369           PU_05         13 hatchery         1
#> 370           PU_05         13     wild         0
#> 371           PU_08         13 hatchery         0
#> 372           PU_08         13     wild         0
#> 373           PU_07         13 hatchery         0
#> 374           PU_07         13     wild         0
#> 375           PU_04         13 hatchery       277
#> 376           PU_04         13     wild        27
#> 377           PU_15         13 hatchery        31
#> 378           PU_15         13     wild         6
#> 379           PU_09         13 hatchery       478
#> 380           PU_09         13     wild       115
#> 381           PU_10         13 hatchery         0
#> 382           PU_10         13     wild         0
#> 383           PU_11         13 hatchery        21
#> 384           PU_11         13     wild         7
#> 385           PU_12         13 hatchery         0
#> 386           PU_12         13     wild         0
#> 387           PU_13         13 hatchery         8
#> 388           PU_13         13     wild         5
#> 389           PU_14         13 hatchery      1790
#> 390           PU_14         13     wild       429
#> 391           PU_01         14 hatchery         0
#> 392           PU_01         14     wild         0
#> 393           PU_02         14 hatchery         1
#> 394           PU_02         14     wild         1
#> 395           PU_03         14 hatchery        10
#> 396           PU_03         14     wild        18
#> 397           PU_06         14 hatchery         5
#> 398           PU_06         14     wild         3
#> 399           PU_05         14 hatchery         0
#> 400           PU_05         14     wild         0
#> 401           PU_08         14 hatchery         0
#> 402           PU_08         14     wild         0
#> 403           PU_07         14 hatchery         0
#> 404           PU_07         14     wild         0
#> 405           PU_04         14 hatchery        49
#> 406           PU_04         14     wild        61
#> 407           PU_15         14 hatchery         6
#> 408           PU_15         14     wild         9
#> 409           PU_09         14 hatchery       135
#> 410           PU_09         14     wild       214
#> 411           PU_10         14 hatchery         0
#> 412           PU_10         14     wild         0
#> 413           PU_11         14 hatchery         6
#> 414           PU_11         14     wild        14
#> 415           PU_12         14 hatchery         0
#> 416           PU_12         14     wild         0
#> 417           PU_13         14 hatchery         2
#> 418           PU_13         14     wild         9
#> 419           PU_14         14 hatchery       494
#> 420           PU_14         14     wild       950
#> 421           PU_01         15 hatchery         0
#> 422           PU_01         15     wild         0
#> 423           PU_02         15 hatchery         1
#> 424           PU_02         15     wild         1
#> 425           PU_03         15 hatchery        21
#> 426           PU_03         15     wild        36
#> 427           PU_06         15 hatchery         8
#> 428           PU_06         15     wild         4
#> 429           PU_05         15 hatchery         0
#> 430           PU_05         15     wild         0
#> 431           PU_08         15 hatchery         0
#> 432           PU_08         15     wild         0
#> 433           PU_07         15 hatchery         0
#> 434           PU_07         15     wild         0
#> 435           PU_04         15 hatchery       118
#> 436           PU_04         15     wild        99
#> 437           PU_15         15 hatchery        13
#> 438           PU_15         15     wild        17
#> 439           PU_09         15 hatchery       224
#> 440           PU_09         15     wild       320
#> 441           PU_10         15 hatchery         0
#> 442           PU_10         15     wild         0
#> 443           PU_11         15 hatchery        11
#> 444           PU_11         15     wild        21
#> 445           PU_12         15 hatchery         0
#> 446           PU_12         15     wild         0
#> 447           PU_13         15 hatchery         4
#> 448           PU_13         15     wild        13
#> 449           PU_14         15 hatchery       878
#> 450           PU_14         15     wild      1362