Run the Dam Impact Analysis for One Generation

Use functions from dia to simulate one generation of 2 sea-winter female Atlantic salmon through time using data and inputs from Nieland et al. (2013, 2015, 2020) as implemented in `@Risk`. Not intended to be called directly. All argument values are inherited from run_dia by default.

Usage

run_one_gen(
  wild_adults,
  hatchery_adults,
  stocking,
  n_stocked,
  upstream,
  downstream,
  in_river_s,
  mattaceunk_impoundment_mortality,
  p_stillwater,
  indirect_latent_mortality,
  p_female,
  new_or_old,
  marine_s_hatchery,
  marine_s_wild,
  straying_matrix,
  p_mainstem_up,
  n_broodstock
)

Arguments

wild_adults

Number of starting wild adult salmon.

hatchery_adults

Number of starting hatchery adult salmon.

stocking

Number indicating whether hatchery stocking is on (`1`) or off (`0`).

n_stocked

Number of smolts stocked in one generation.

upstream

Upstream passage efficiency of adult Atlantic salmon through dams.

downstream

Downstream survival of Atlantic salmon smolts through dams. The default values are based on the most recent values used by Nieland and Sheehan (2020) based on standards established in the the NOAA Species Protection Plan following implementation of the Penobscot River Restoration Project. The default `NULL` randomly samples correlated survival rates of smolts for each dam based on cumulative flow probabilities and associated empirical survival rates (Nieland et al. 2013, Nieland and Sheehan 2020).

in_river_s

In-river survival per kilometer for downstream migrating smolts. The default `NA` value simulates from cumulative distribution function using values in in_river_m.

mattaceunk_impoundment_mortality

Mortality incurred by Atlantic salmon smolts during migration through the Mattaceunk (Weldon) Dam impoundment. The default value is based on Nieland and Sheehan (2020). Based on results of Holbrook et al. (2011) and Stich et al. (2015a).

p_stillwater

Probability that fish use the Stillwater Branch for downstream migration. The default (`NA`) draws flow-correlated values from a cumulative distribution of flows with paired estimates of p_stillwater used by Nieland and Sheehan (2020), based on empirical results in Holbrook et al. (2006), Stich et al. (2014), and Stich et al. (2015a).

indirect_latent_mortality

Indirect, latent mortality incurred by Atlantic salmon smolts at each dam passed. The default value is what was used in Nieland and Sheehan (2020), derived from estimates in Stich et al. (2015b).

p_female

Proportion of females in population.

new_or_old

A character string indicating whether to use `"new"` (Nieland et al. 2013, 2015) or `"old"` (Nieland and Sheehan 2020) flow-correlated probabilities of p_stillwater as well as flow-correlated survival at `milford`, `orono`, and `stillwater` dams.

marine_s_hatchery

Numeric indicating marine survival rate for post-smolt to adult survival of hatchery outmigrants. The default (`NULL`) simulates values from a truncated normal distribution using hatchery smolt survival estimates from the Penobscot River, ME, USA.

marine_s_wild

Numeric indicating marine survival rate for post-smolt to adult survival of wild outmigrants. The default (`NULL`) simulates values from a truncated normal distribution using wild smolt survival estimates from the Narraguagus River, ME, USA.

straying_matrix

A dataframe identical in structure to the built-in straying_locations dataset.

p_mainstem_up

Probability that fish use the mainstem Penobscot River (and not Stillwater Branch) for upstream migration around Marsh Island. The default value (`NULL`) draws flow-correlated probabilities based on Holbrook et al. (2009).

n_broodstock

Target number of adult returns collected at Milford Dam for spawning at US Fish and Wildlife Service Craig Brook National Fish Hatchery each year. Broodstock are collected upstream of Milford Dam in run_upstream_passage.

Value

A dataframe with 15 observations of 3 variables.

production_unit Production unit hatchery_adults Number of hatchery adults in each PU wild_adults Number of wild adults in each PU

References

Holbrook CM, Kinnison MT, Zydlewski J. 2011. Survival of migrating Atlantic salmon smolts through the Penobscot River, Maine: a prerestoration assessment. Transactions of the American Fisheries Society 140:1255–1268.

Holbrook CM, Zydlewski J, Gorsky D, Shepard SL, Kinnison MT. 2009. Movements of prespawn adult Atlantic salmon near hydroelectric dams in the lower Penobscot River, Maine. North American Journal of Fisheries Management 29:495–505.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams on diadromous fish: a case study for Atlantic salmon in the Penobscot River, Maine. ICES Journal of Marine Science 72:2423–2437.

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US Department of Commerce, Northeast Fisheries Science Center Reference Document 13-09, Woods Hole, MA.

Stich DS, Bailey MM, Holbrook CM, Kinnison MT, Zydlewski JD. 2015a. Catchment-wide survival of wild- and hatchery-reared Atlantic salmon smolts in a changing system. Canadian Journal of Fisheries and Aquatic Sciences 72:1352–1365. Stich DS, Bailey MM, Zydlewski JD. 2014. Survival of Atlantic salmon Salmo salar smolts through a hydropower complex. Journal of Fish Biology 85:1074–1096. Stich DS, Kinnison MT, Kocik JF, Zydlewski JD. 2015b. Initiation of migration and movement rates of Atlantic salmon smolts in fresh water. Canadian Journal of Fisheries and Aquatic Sciences 72:1339–1351.